fix: extract function to utils
